test: . Move MockDataLoader to same directory as MockDataHelpers.ts
