+++ tests/TestingTools/MockDataLoader.test.ts
@@ -1,6 +1,5 @@
+import one_task from '../Obsidian/__test_data__/one_task.json';
 import { MockDataLoader } from './MockDataLoader';
-
-import one_task from './__test_data__/one_task.json';
 
 describe('MockDataLoader', () => {
     it('should return data for a known file', () => {
